% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateF7Login}
\alias{updateF7Login}
\title{Toggle login page}
\usage{
updateF7Login(id, user = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{\link{f7Login} unique id.}

\item{user}{Value of the user input.}

\item{session}{Shiny session object.}
}
\description{
Toggle login page
}
