% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createDBCon}
\alias{createDBCon}
\title{Create connection to a Postgres database.}
\usage{
createDBCon()
}
\description{
Parameters are extracted from Golem options
which is convenient to keep credentials secrets
}
