% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_grid.R
\name{grid_init}
\alias{grid_init}
\alias{grid_gen_mines}
\alias{grid_numbers}
\alias{grid_data}
\alias{generate_spatial_grid}
\title{Grid generation}
\usage{
grid_init(N)

grid_gen_mines(grid, n_mines)

grid_numbers(grid)

grid_data(grid, N)

generate_spatial_grid(N, n_mines)
}
\arguments{
\item{N}{size of the grid}

\item{grid}{the grid}

\item{n_mines}{number of mines}
}
\description{
Several functions to generate the minesweeper grid.
}
