% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tests.R
\name{html2R}
\alias{html2R}
\title{Convert HTML content to R Shiny tags}
\usage{
html2R(htmlStr, prefix = FALSE)
}
\arguments{
\item{htmlStr}{HTML string}

\item{prefix}{Whether to prefix elements by tag$...}
}
\value{
A list of R Shiny tags
}
\description{
Convert HTML content to R Shiny tags
}
\author{
Alan Dipert, RStudio
}
