% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bomb_counter.R
\name{mod_bomb_counter_ui}
\alias{mod_bomb_counter_ui}
\alias{mod_bomb_counter_server}
\title{mod_bomb_counter_ui and mod_bomb_counter_server}
\usage{
mod_bomb_counter_ui(id)

mod_bomb_counter_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{r}{cross module variable}
}
\description{
A shiny Module.
}
\keyword{internal}
