% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chat.R
\name{mod_chat_ui}
\alias{mod_chat_ui}
\alias{mod_chat_server}
\title{mod_chat_ui and mod_chat_server}
\usage{
mod_chat_ui(id)

mod_chat_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{r}{cross module variable}
}
\description{
A shiny Module.
}
\keyword{internal}
