% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_scores.R
\name{mod_display_scores_ui}
\alias{mod_display_scores_ui}
\alias{mod_display_scores_server}
\title{mod_display_scores_ui and mod_display_scores_server}
\usage{
mod_display_scores_ui(id)

mod_display_scores_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{r}{cross module variable}
}
\description{
A shiny Module.
}
\keyword{internal}
