% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_game_grid.R
\name{mod_game_grid_ui}
\alias{mod_game_grid_ui}
\alias{mod_game_grid_server}
\title{mod_game_grid_ui and mod_game_grid_server}
\usage{
mod_game_grid_ui(id)

mod_game_grid_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
