% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_help.R
\name{mod_help_ui}
\alias{mod_help_ui}
\alias{mod_help_server}
\title{mod_help_ui and mod_help_server}
\usage{
mod_help_ui(id)

mod_help_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
