% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_share.R
\name{mod_share_ui}
\alias{mod_share_ui}
\alias{mod_share_server}
\title{mod_share_ui and mod_share_server}
\usage{
mod_share_ui(id)

mod_share_server(id, r)
}
\arguments{
\item{id}{shiny id}

\item{r}{cross module variable}
}
\description{
A shiny Module.
}
\keyword{internal}
