% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timer.R
\name{mod_timer_ui}
\alias{mod_timer_ui}
\alias{mod_timer_server}
\title{mod_timer_ui and mod_timer_server}
\usage{
mod_timer_ui(id)

mod_timer_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{r}{cross module variable}
}
\description{
A shiny Module.
}
\keyword{internal}
