% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_welcome.R
\name{mod_welcome_ui}
\alias{mod_welcome_ui}
\alias{mod_welcome_server}
\title{mod_welcome_ui and mod_welcome_server}
\usage{
mod_welcome_ui(id)

mod_welcome_server(id, r)
}
\arguments{
\item{id}{shiny id}

\item{r}{cross module variable}
}
\description{
A shiny Module.
}
\keyword{internal}
