% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\name{cases_to_reveal}
\alias{cases_to_reveal}
\alias{reveal_onclick}
\title{Create a list of cases to reveal}
\usage{
cases_to_reveal(res, case_id, N)

reveal_onclick(data, case_id, N)
}
\arguments{
\item{res}{Leaflet data.}

\item{case_id}{Case to reveal.}

\item{N}{Grid size.}

\item{data}{Leaflet data.}
}
\value{
List containing cases to reveal.
}
\description{
This function is used by \link{reveal_onclick}
}
