% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  usecase = "local",
  dbname = "",
  dbhost = "",
  dbport = "",
  dbuser = "",
  dbpwd = "",
  table_scores = "",
  table_message = "",
  avatars = list.files("inst/app/www/avatars/")
)
}
\arguments{
\item{usecase}{usecase}

\item{dbname}{database name}

\item{dbhost}{database host}

\item{dbport}{database port}

\item{dbuser}{database user}

\item{dbpwd}{database password}

\item{table_scores}{database table name}

\item{table_message}{database table name}

\item{avatars}{List of avatar images. Needed by \link{generateAvatar} in 
\link{mod_display_scores_server}.}
}
\description{
Run the Shiny Application
}
