% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_nickname}
\alias{validate_nickname}
\title{Validate a nickname provided by the user.}
\usage{
validate_nickname(char)
}
\arguments{
\item{char}{Nickname.}
}
\value{
Boolean
}
\description{
Used by the \link{mod_welcome} module
}
