% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_game_params.R
\name{mod_game_params_ui}
\alias{mod_game_params_ui}
\alias{mod_game_params_server}
\title{mod_game_params_ui and mod_game_params_server}
\usage{
mod_game_params_ui(id)

mod_game_params_server(id, r)
}
\arguments{
\item{id}{shiny id}

\item{r}{cross module variable}
}
\description{
A shiny Module.
}
\keyword{internal}
