% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  usecase = "local",
  dbname = "",
  dbhost = "",
  dbport = "",
  dbuser = "",
  dbpwd = "",
  table_name = ""
)
}
\arguments{
\item{usecase}{usecase}

\item{dbname}{database name}

\item{dbhost}{database host}

\item{dbport}{database port}

\item{dbuser}{database user}

\item{dbpwd}{database password}

\item{table_name}{database table name}
}
\description{
Run the Shiny Application
}
