% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(usecase = "local", ec_host = "", ec_room = "")
}
\arguments{
\item{usecase}{usecase}

\item{ec_host}{ethercalc host}

\item{ec_room}{ethercalc room}
}
\description{
Run the Shiny Application
}
